more normalizers in the correct way
